core-thrash: remove ptrace as this can cause lockups
Attaching to all processes with ptrace is not adviseable as it can
stop terminals/shells/X etc and then stress-ng is not easily killable,
so remove this.  Add more yield points too and localised SIGALRM yield
checking to make the thrashing process more responsive to termination.
Also increase the page thrashing probability on each thrash loop
to increase paging in for more thrashing.

Signed-off-by: Colin Ian King <colin.king@canonical.com>