Add memory barriers an ready flag to check if counter is in a sane state
The counter now is protected with a ready flag that is set to false
during the counter update.  The ordering of the flag unset, counter
update and flag set is protected with memory barriers to enforce
strict ordering.  Add a check at the end of the run to check if
the counter ready flag is in a sane state and fail with an error
if it is set incorrectly.  Addresses concerns raised by ELISA.

Signed-off-by: Colin Ian King <colin.king@canonical.com>