stress-vm: fix bit error checking when bogo-ops limits reached
There are some vm methods that abort early when the bogo-ops limit
has been reached and then sanity check the all the data even though
the all the data has not been set because of the premature abort.
Fix this so that we only sanity check full runs of data.

Signed-off-by: Colin Ian King <colin.king@canonical.com>