stress-enosys: prevent segfault when doing syscall_find
syscall_find and syscall_add will index the hash table using the division
remainder of number, which may be negative. This will cause the stressor to
stop running when testing syscalls with high bits as the most significant
bit is non-zero. Using unsigned longs for the indexes fix the problem.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>