stress-{iomix,fiemap}: fix copying of args struct
The current args copying is ommitting the new counter_ready pointer
and this causes segfaults. Fix this by copying the args counter using
structure assignment to ensure all the fields are copied and then
make the necessary local adjustments. Requires relating the constness
of the counter and counter_ready pointers.

Signed-off-by: Colin Ian King <colin.king@canonical.com>