core-sched: add new sched settings helper sched_settings_apply
This helper function fetches the required settings and applies (and/or
re-applies them). This cleans up some duplication of code with the
deadline scheduler setting.  The single argument "quiet" when true
will try to set the settings and won't report any failures, this is
useful in a child context when we don't want to spam the output. In
the parent context, setting quiet to false is useful as this will
report issues when the user tries to use incorrect setting parameters.

Kudos to Chunyu-Hu for the original idea and a suggested fix for the
issue with respect to deadline scheduling, I based my code on this
idea and re-worked the code for all cases.

Signed-off-by: Colin Ian King <colin.king@canonical.com>