core-sched: init shim_sched_attr with zero
stress-ng defined the sched_attr struct, if the version is different from
the kernel side, an '-E2BIG' would be returned to user by kernel, iff
the size of the user defined version is larger than kernel and the larger
part is not zero (man sched_setattr). So init it to avoid the issue.

Signed-off-by: Chunyu Hu <chuhu@redhat.com>