stress-sysinval: add MISC args allowing finer system call argument passing
As well as using generic argument types, add a MISC type that is not based
on a mixed bitmask of options but instead one of many possible misc args
for system calls.

Also check clock timeout per permute call as we can get stuck deep in
system calls and this can take a while for the child to reach the permute
last arg before we check the timeout.

Signed-off-by: Colin Ian King <colin.king@canonical.com>