stress-get: handle ENOSYS from shim wrappers as no fatal, remove verify check
The shim wrappers may return ENOSYS of the __NR_* defines are not
defined for some architectures, so don't throw a fatal error if ENOSYS
is returned.  Also remove the verify check as the check is so light
weight it is more expensive to check verify is set than just do the
check by default.

Signed-off-by: Colin Ian King <colin.king@canonical.com>