stress-rlimit: fix segv caused by passing a NULL context and exercise all rlimit resource ids
Currently the context is being passed to the child as a NULL pointer causing
it to segfault prematurely. Fix this by passing the address of the context
structure.  Also exercise all the known rlimit resource ids by getting and
setting them.  Finally, exercise the fetching of an illegal resource id.

Signed-off-by: Colin Ian King <colin.king@canonical.com>