stress-ng.h: Add a SHIM_CAP_IS_ROOT shim capability
Add SHIM_CAP_IS_ROOT so that stress_check_capability can check
for root privilege if a capability is not supported. This is
currently 0 but this clashes with CAP_CHOWN. The next change
will replace all stress_check_capability using 0 with SHIM_CAP_IS_ROOT
and then change SHIM_CAP_IS_ROOT to be non-zero and not clashing
with any CAP_* values.

Signed-off-by: Colin Ian King <colin.king@canonical.com>