core-helper: add stress_get_max_file_limit to return max file open limit
Add a new helper to return the maximum number of files a process
can open ignoring the number of already opened file descriptors. This
is faster and less accurate than stress_get_file_limit but is suitable
for some stress tests that can handle ENFILE or EMFILE correctly when
we run out of file descriptors.

Signed-off-by: Colin Ian King <colin.king@canonical.com>