stress-zlib: Fixes race conditions during deflate/inflate
1. Parent process deflates data and writes it to a pipe. This could be
interrupted by alarm and then lead to an EINTR error. But this should be
handled properly and make sure that everything was written to the pipe.

2. The functions read() and write() can return with less than size bytes
read or written. Therefore we have to take care this will be handled
properly. Otherwise we could end up reading less bytes than required to
distinquish the buffer from the next size value in the pipe.

Signed-off-by: André Wild <wild1@linux.ibm.com>