stress-memthrash: fix premature termination by OOM handler
Currently the OOM handler parent gets a SIGALRM that causes it to
kill the child and the pthreads can continue running updating the
bogo-ops counter because the child cannot wait for them to complete.

Relax the OOM handler to now send several SIGARLRMs to the child
process before escalating it to SIGTERM and SIGKILL and add a SIGALRM
handler in the child to perform pthread reaping.

Signed-off-by: Colin Ian King <colin.king@canonical.com>