stress-dev: add some device timeouts on slow optical devices
The ioctls being exercised can block for a long period of time
on slow optical devices, so add an itimer to cause a -EINTR on
interruptible ioctls.  For long duration ioctls, check how
long has been taken and mark these to be ignore on the next
iteration so to get more test throughput.

Signed-off-by: Colin Ian King <colin.king@canonical.com>