stress-fanotify: add test_fanotify_mark
The files were mounted on each iteration of the stressor loop,
slowing down the stress test with these calls. Mount the files
at the start of the stressor and passed it as an argument to the
fanotify_event_init function. Add function test_fanotify_mark
to test and increase kernel coverage of fanotify_mark syscall.

Signed-off-by: Piyush Goyal <piyushgoyaliit@gmail.com>