stress-af-alg: check for ELIBBAD on bind
On FIBS enabled kernels the bind to the af-alg socket can fail with
-ELIBBAD if the algorithm was not allowed by FIPS. In these cases
the /proc/crypto info marks these as not passed on the self tests,
so we can check this state when ELIBBAD errors occur. Only throw
an error if the self test passed (e.g. the algo was allowed but
the bind failed for some other reason).

Signed-off-by: Colin Ian King <colin.king@canonical.com>