stress-vdso: fix symbol lookup on clang built kernels
I made an assumption that the hash table and string tables were
always defined before the symbol table, but it appears that the
ordering isn't strict and VDSO's generated with clang builds and
linked using lld.  Fix this by scanning for the hash table and
string tables and then scanning for the symbol table.

Fixes: https://github.com/ClangBuiltLinux/linux/issues/1175

Signed-off-by: Colin Ian King <colin.king@canonical.com>