stress-tmpfs: fix incorrect lseek failure check
Currently the return from large lseeks was resulting in failures
being incorrectly detected when in fact the return was just too large
for an int.  Fix this.  Also don't make overly large tmpfs files
as this slows down the test on massive memory systems, instead set
a maximim limit of 512MB.  The original 98% of memory of more than
512MB can be enabled using the --maximize option.

Signed-off-by: Colin Ian King <colin.king@canonical.com>