stress-session: improve error handling from child process
It is useful to pass the errno back to the parent to report more
failure information especially if the child dies prematurely. Add
in a communications pipe to return the status and errno of any
failures. Also add a new error STRESS_SESSION_WRONGSID_FAILED to
report incorrect session id failures.

Signed-off-by: Colin Ian King <colin.king@canonical.com>