stress-cpu: fix --cpu-load timer issues
The --cpu-load delays are a bit broken. Fix this by try to measure
CPU time used by the CLOCK_PROCESS_CPUTIME_ID timer, but if this
is not available fall back to using the wall clock time. When the
allocated CPU slice is complete, ensure the delay sleep is based
on wall clock time - don't use the CLOCK_PROCESS_CPUTIME_ID timer
for this delay measurement as it measures CPU time used and that
will be negligable for a sleeping delay.

Signed-off-by: Colin Ian King <colin.king@canonical.com>