stress-aio-linux: ensure checking is really correct
The original checking code was fundamentally broken. Fix the
write and read offsets so they match.  Improve the io waiting
to keep track of the number of waits that got completed.  Remove
the aync sync'ing as this can flush data back onto the file during
the next write/read test cycle and this messes up the checking.

Finally, the check data values are now based on the buffer address
shifted by the page size + an incrementing value to ensure we match
buffer contents with the buffer.

Signed-off-by: Colin Ian King <colin.king@canonical.com>