Strengthen verification with --verify --hdd N
Today when running with `--verify --hdd N` the verification is
effectively only checking that every 512th byte read from the file is a
`0`. This risks missing bugs in which writes are missed, pages are
zeroed, or writes go to the wrong pages.

This commit verifies the content of the file more thoroughly, and adds
more variety to the bytes that are verified to reduce the risk of
missing such bugs.