Add new secretmem stress test
Introduce a simple memfd_secret secret memory allocation stress
test. This tries to create lots of secret memory mappings and will
try to fragment the mappings.

Each stress loop iteration will expand the mappable region by 3
pages using ftruncate and mmap and touches the pages. The pages are then
fragmented by unmapping the middle page and then umapping the first and
last pages. This tries to force page fragmentation and also trigger out of
memory (OOM) kills of the stressor when the secret memory is exhausted.
Note this is a Linux 5.11+ only stressor and the kernel needs to be booted
with "secretmem=" option to allocate a secret memory reservation.

Signed-off-by: Colin Ian King <colin.king@canonical.com>