core-try-open: resolve some stress-dev blocking issues, add richer error return
Currently the fork/wait of a child that performs the open can end up
with multiple zombies.  A fix to this is to defer the open to a grandchild
process that can be reaped by the child or init if we need to kill the
child.  Improve the helper function by adding some additional error
return status types so we can determine if an open failed because of
an open failure or because of a fork/wait or exit wait status failed.

Signed-off-by: Colin Ian King <colin.king@canonical.com>