stress-dev: only perform open tries if never opened before
The open try mechanism is expensive in fork/wait overhead, so only
perform this on the first time a device is tested to see if it
can be opened without hard locking.  Also limit the number of open
locked devices so that we don't end up using up all the process
slots on blocked child processes that are in a zombie state.

Signed-off-by: Colin Ian King <colin.king@canonical.com>