stress-resources: fix mq files not being reaped
There are still a couple more issues to be resolved to fix the
mq files not being properly repead. Firsly remove a yield point
from the resource reaping loop, this was breaking too early and
all the resources were not being cleaned up. Also force remove
the files even if the mq_open failed if the file name is available
in case mq_open created a file and the mq_open was prematurely
interrupted (however unlikely that is).

Signed-off-by: Colin Ian King <colin.king@canonical.com>