stress-cache: ignore sfence if not available, add missing help and optimize
The sfence option should be ignored if sfence is not supported by
the architecture, so disable it if not available. There is a missing
cache-no-affinity help information, so add it and re-order the
help information too. Finally, optimize the cache wrapping index
by removing modulo and using a ternary operator that is fast more
efficient.

Signed-off-by: Colin Ian King <colin.king@canonical.com>