Add a STRESS_TRY_AGAIN don't give up flag and simplify try-open
Some open failures occur because a device is busy or memory is slow,
so mark these as a STRESS_TRY_AGAIN error to allow caller to decide
to try again or silently move on. Simplify the open blocking, add
a alarm(1) call to nudge blocked opens with a EINTR failure. Also
don't SIGKILL the child but send SIGALRM to trigger timeout cleanups

Signed-off-by: Colin Ian King <colin.king@canonical.com>