stress-pthread: clean up cppcheck warning and use coding style for multiple #defines
cppcheck complains if the two args of the STRESS_MAXIMUM macro are the
same. Work around this by defining a default stack minimum macro and only
checking the minimum of this macro is not the same as PTHREAD_STACK_MIN.

Also break long #if defined() checks into multiple lines to match stress-ng
coding style.

Signed-off-by: Colin Ian King <colin.king@canonical.com>