stress-cache: optimize out switch statement and % operators
The switch statement impacts instruction branching because it
has to scan down the 16 different cases. Optimize this by making
each case a small static function that performs the cache exercising
loop and call the appropriate function based on a lookup table.

Also make the step increnements less than the cache size to remove
the % operator and just use an wrap around check using the ternary
operator; this compiles into far faster code as % is really slow.

Signed-off-by: Colin Ian King <colin.king@canonical.com>