stress-l1cache: add new level 1 cache stressor
start N workers that exercise the CPU level 1 cache with reads
and writes. A cache aligned buffer that is twice the level 1
cache size is read and then written in level 1 cache set sized
steps over each level 1 cache set. This is designed to exercise
cache block evictions. The bogo-op count measures the number of
million cache lines touched. Where possible, the level 1 cache
geometry is determined from the kernel, however, this is not
possible on some architectures or kernels, so one may need to
specify these manually. One can specify 3 out of the 4 cache
geometric parameters, these are as follows:

	--l1cache-line-size N
	--l1cache-sets N
	--l1cache-size N
	--l1cache-ways N

Signed-off-by: Colin Ian King <colin.king@canonical.com>