stress-fork: done use fork shim helper, move code into core stressor loop
Having a shim wrapper for fork isn't such a bright idea as it can
cause issues on some architectures.  Move this into the main stressor loop.
Also decide which fork or vfork to call based on a which argument rather
than pass the function indirectly to avoid any indirection calling too.
Finally, no need for the parent to kill the child as the child will call
_exit.  Add a yield call in the child to force extra scheduling and make
the stressor more yieldy at the cost of more context switches

Signed-off-by: Colin Ian King <colin.king@canonical.com>