Makefile: pass LDFLAGS to function test builds, echo configs on VERBOSE=1
Currently the required libraries are passed to the function test build
checks causing some functions to be falsely detected as available when
they can't be used in static builds.  Fix this by passing the detected
CONFIG_LDFLAGS setting that has the detected usable libraries to the
test build so just these are used instead.  Also when in VERBOSE build
mode echo out the CONFIG_LDFLAGS and CONFIG_CFLAGS to enable easier
debugging if configs fail.

Signed-off-by: Colin Ian King <colin.king@canonical.com>