stress-sigabrt: use handler and non-handler modes of SIGABRT
Calling abort() with a handler installed will cause the signal handler
to be called, then abort will disable the handler, re-signal and cause
the process to exit.  Raising SIGABRT without a handler will terminate
the process.  Exercise these two modes and add some sanity checks to
see if this is working as per the abort() and SIGABRT POSIX specification.

Signed-off-by: Colin Ian King <colin.king@canonical.com>