stress-cpu: increase number of iterations for omega stressor
BugLink: https://github.com/ColinIanKing/stress-ng/issues/120

The omega stressor requires one more iteration for Solaris x86 since
the precision in the math ops are slightly different to that of x86
glib on other platforms. Add one more iteration.

Kudos to achelon5 for reporting and helping to debug this issue.

Signed-off-by: Colin Ian King <colin.king@canonical.com>