stress-sigio: fix segfault (LP: #1918095)
The SIGIO handler was killing the child pid and this was leading
to segfaults when pending I/O (e.g. pr_inf) occurred and buffers
were being flushed.  Clean up the code to only perform the signal
handler in the parent process and not the child, and remove the
process killer in the signal handler.

Signed-off-by: Colin Ian King <colin.king@canonical.com>