stress-dentry: add yield point detection before starting iteration
Currently the yield point check occurs after a directory entry create
has occurred leaving the end count off by one so that files are
potentially not removed if a yield interrupt hits after a mknod
occurs. Move the yield point detection to the start of the loop
to fix this.

Signed-off-by: Colin Ian King <colin.king@canonical.com>