stress-io-uring: mmap buffers with MAP_POPULATE
Don't use the heap for buffer allocation, use mmap since this is
page aligned (and hence block aligned) and can be populated before
use. This may be wasteful for 512 byte block sizes but it does also
protect the heap a little better from being stompted on if things go
haywire.

Signed-off-by: Colin Ian King <colin.king@canonical.com>