stress-zlib: fix incorrect buffer size on compression
The buffer size has been 4 times smaller than it should due to
a mismatch between the type of the buffer (uint32_t elements) and
the size. The size was assumed to be number of these elements but
it was the size of the entire buffer. Fix this and make the buffer
type uint8_t and re-work all the data generators to use the size
as the number of uint8_t elements in the buffer.

Signed-off-by: Colin Ian King <colin.king@canonical.com>