stress-bad-altstack: Fix hang on OS X
If the child ends up in a recursive segfault on OS X then the child never
terminates and the parent can wait forever using the shim_waitpid variant
of waiting. Replace this with the normal waitpid call as the bad-altstack
stressor handles EINTR for the SIGALRM or SIGINT style of terminations.

Signed-off-by: Colin Ian King <colin.king@canonical.com>