stress-vforkmany: fix handling on OS X where vfork() fails
vfork may fail on OS X because it does not have an exec or exit following
the call, so cater for this failure case. Fix the started logic to just
the pid rather than a flag, this makes it easier to track the starting
parent pid.  Finally remove the vfork shim as this leads to botched stack
return paths when the code is not inlined.

Signed-off-by: Colin Ian King <colin.king@canonical.com>