stress-sysfs: catch killing signals when exercising sysfs
Some sysfs files on buggy kernels have been observed to kill the
sysfs stressor. This causes a bogo-ops counter reporting that the
stressor terminated badly.  To avoid this, run the sysfs stressor
as a child and have a parent monitor it exiting. If it was killed
by an unexpected signal from the kernel then report this and the
most recent sysfs file that was being exercising as a debug message
(run with -v option) to help diagnose why it was killed.

This involves putting the context into a shared mmapping so that
the parent and child can track state between them.

Signed-off-by: Colin Ian King <colin.king@canonical.com>