stress-sock: fix recvmmsg infinite loop
The recvmmsg size check currently checks the number of msgvec messages
handled and assumes that non-zero means more data can be read. In fact
we need to count the size of the msg_len fields to see if data has been
read with zero indicating no more data and time to exit the read loop.

Signed-off-by: Colin Ian King <colin.king@canonical.com>