Add --skip-silent option to silent skip stress tests that can't be invoked
silence messages that report that a stressor has been skipped because it
requires features not supported by the system, such as unimplemented system
calls, missing resources or processor specific features.

Add two logging helpers pr_inf_skip and pr_dbg_skip to check for the silence
flag and not report messages if the flag is enabled.

Signed-off-by: Colin Ian King <colin.king@canonical.com>