Fix duration output missing when sequential mode
When using the '--seq' option, stress-ng will output the following line:

stress-ng: info:  [25541] defaulting to a 60 second run per stressor

it means the test duration per stressor is the 60s

However, when using the -t option, this information is not shown.

Fix this by displaying the duration for each stressor.

[ This is a minor re-working of the original fix by James Wang and
  a minor rewording of the commit message ]

Signed-off-by: James Wang <jnwang@linux.alibaba.com>
Signed-off-by: Colin Ian King <colin.king@canonical.com>