stress-affinity: add new --affinity-pin and --affinity-delay options
The new --affinity-pin option will pin the child processes of each
stressor instance to a CPU in sync as it walks through the available
CPUs.  The new --affinity-delay allows one to burn cycles pegged on
a CPU performing scheduling yields for a specified time in nanoseconds.

A typical usecase may be:

stress-ng --affinity 1 --affinity-pin --affinity-delay 500000000

This change CPU affinity every 0.5 seconds and ensure all the 16 sub
child affinity worker processes are all using the same CPU.

Or for say a 256 CPU system:

stress-ng --affinity 64 --affinity-pin --affinity-delay 1000000000

This will cause 64 workers to pin onto 25% of the 256 processors
and change cpu affinity every second.

Signed-off-by: Colin Ian King <colin.king@canonical.com>