stress-resources: reap pthread only if it was started
The pthread join will fault if the pthread was not started because
the initial pthread return setting is zero which so happens to be
the pthread success return value. Fix this by initializing it with
-1 and also initialize the pthread thread handle to 0 and checking
for this before joining.

Kudos to Vincent Chen for the work in debugging and diagnosing this
bug.

Signed-off-by: Colin Ian King <colin.king@canonical.com>