stress-dirmany: Add a stressor to rapidly create/remove empty files
Some file systems (such as ext4) have shown issues in the past when
populating a directory with hundreds of thousdands of files. This
new stressor will create and remove as many files as possible in
a directory. It uses a rule-of-thumb that 60% of the time will be
used in file creation and 40% in file removal to try and create
as many files and remove them in a timely manner.

Signed-off-by: Colin Ian King <colin.king@canonical.com>