Add file hole punching (and filling) stressor, --fpunch
This exercises a file by punching and filling holes in the file
with 5 concurrent processes per stressor hammering the same file
generating multiple extents.  The offsets and sizes of holes being
generated and filled may fail for some of the fallocate calls and
we silently ignore these. The aim of the stressor is to try and
break fallocated extents with lots of concurrent racy actions.

Signed-off-by: Colin Ian King <colin.king@canonical.com>