stress-sigio: drain as much as possible in SIGIO handler
Ensure that the SIGIO handler drains the buffer rather
than doing a single read. This fixes an assumption that
SIGIO occurs on each write. Also make the buffer sizes the
same and tweak the pipe size to allow two buffer's worth
of data in the pipe.

Add some SIGIO rate stats too.

Signed-off-by: Colin Ian King <colin.king@canonical.com>