core-vmstat: fix trim of devnames for nvme devices
In order to obtain I/O statistics with '--iostat' the device name of the
full disk must be derived from the partition device name.

Currently this is done by cutting of the partition number at the end of
the device name e.g. /dev/sda1 -> /dev/sda.

However, for nvme devices there is an additional 'p' character serving as
delimiter between the disk name and the partition number, e.g.
/dev/nvme0n1p1.

The result is that running stress-ng with '--iostat' from a directory
located on a nvme disk is not giving proper output, since nvme{N}n{M}p
is not a valid disk name.

Fix this by adding another trim function to cut of the trailing 'p'
character after the digits have been trimmed, e.g. /dev/nvme0n1p ->
/dev/nvme0n1.

Signed-off-by: Danilo Krummrich <danilokrummrich@dk-develop.de>