core-helper: Quote the system-info.version value in YAML metrics
The value of `system-info.version` typically begins with a pound sign.
When used as a bare value, any YAML parser will interpret the value as
`null` and followed by a comment.

This change has the value quoted in the YAML, so the pound sign is
actually considered part of the value.