stress-ng: Fix --temp-path directory checking (#163)
Only check if the temp path is accessible after the options have been
parsed.  Add a helper function to perform the check. Also add a strdup
failure check.  Fixes the case where --temp-path is accessible and
stress-ng is run from a directory that cannot be accessed and temp path
reports a failure because it checks "." first.

Fixes https://github.com/ColinIanKing/stress-ng/issues/163

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>