stress-ng.h: Define the nop instruction for every arch
Although every asm instructions are tested to compile before using them
it is not always sufficient. I've encountered a case where the nop insn
did compile in the test-asm-nop but failed when I have tried to compile
the stress-nop file. In the latter the nop insn is duplicated 64 times,
which didn't compiled for my target architecture -kvx- because the insn
syntax doesn't match. The Kalray VLIW core architecture is kind of like
a DSP, where multiple "instructions" can be executed in the same cycle,
what we call a bundle, which must end with `;;` in the assembly source.

I don't think that simply adding a `;` at the end of every inlined asm
is a good option, it might solve my exact issue but every asm language
have some quirks and, although are very resembling, cannot be expected
to always have the exact same syntax.

I was initially going to let every architecture define it's own set of
instructions strings to be used in both by the "features test", and by
the stressors. But you already have made some changes on that regards,
so this patch ends-up a bit silly without adding the architecture that
required this modification.

Signed-off-by: Jules Maselbas <jmaselbas@kalray.eu>