stress-cache: add more SIGSEGV detection and fix ret return
Currently the sigsetjmp is using variable ret to check for signal
handler jmps, but this is clobbering the return value for ret at
the end of the stressor. Instead, use jmpret so we don't clobber
ret. Also, add two more setjmp calls, one before the signal handler
is installed so we don't siglongjmp back to a garbage address, and
add another one early in the cache loop to catch and respin on
bad cache ops that may have triggered a SIGSEV.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>