stress-bad-altstack: use valid smallest alternative stack
the smallest alternative stack was not correctly set; fix this by
trying to set the alternative stack size until we get a minimum
allowed size for the smallest stack. Also use a global null pointer
to fool static analysis warnings when dereferencing a null pointer
to cause the initial segv.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>