stress-nop: inline each nop instruction and add Kalray VLIW nop
Not all arch nops are the same. The Kalray VLIW core architecture
is kind of like a DSP, where multiple "instructions" can be
executed in the same cycle, what we call a bundle, which must end
with `;;` in the assembly source. To facilitate this different nop
and to handle potentially different arch delay instructions, make
each nop/yield/pause instruction an inlined function.  Add the
Kalray KVX arch detection macro and use this in the nop inline
function for the two different implementations.

Thanks to Jules Maselbas for the original bug report and patches,
without these I would have not been able to model these changes
based on the KVX details provided in Jule's contributions.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>