stress-dev: remove hashing, replace with device list and device state
The hash tracking of device names was sub-optimal. Replace it by
scanning for all the potential device names and storing this information
in a linked list.  The mutable state data that is associated with each
device is now stored in a shared memory mapped region, with a direct
1-to-1 relationship from the device name to the device state. If a
child process terminates then the shared data state is not lost and
can be re-used for further invocations of the child stressors.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>