stress-ng: Fix SIGINT sending alarm signal to all stressors
The process group id and not the -ve pid should be send the
alarm signal. This fixes an issue where SIGINT can prematurely
end up causing the bogo ops to be out of sync with the actual
values because the child is send the signal and not the entire
process group.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>