Add the msyncmany stressor
The stressor mmap's up to 32768 pages onto the first page of a temporary
file so we have multiple mmap'd pages on the same file page. The data is
modified in the first of the mmap'd pages and msync is called, causing
the data to by sync'd back and causing reads on the other pages to be
in sync with the modified page. This exercise msync and page dirtying.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>