stress-rawsock: make client wait for server to start
Add a simple shared page between client and server processes to
allow parent to indicate it is ready to start handling client packets.

This addresses a soft lockup issue as reported in Linug Bugzilla
https://bugzilla.kernel.org/show_bug.cgi?id=215888 and originally
reported and debugged by Thinh Tran. Kudos to Thinh Tran for the
analysis of the issue and a couple of suggested workaround fixes.

Closes https://github.com/ColinIanKing/stress-ng/issues/187

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>