stress-gpu: Add gpu stressor
Add gpu stressor. This stressor uses the gbm API to allocate memory and initializes an opengl context
for stressing the gpu. The shader core usage per pixel and memory bandwidth can be modified by the
gpu options. The default frame buffer sizes x,y is 256 pixels, we can also specify options for
the stressor to perform batches of texture transfers.

Signed-off-by: Dorinda Bassey <dbassey@redhat.com>