[Fix] Clear Hugepages bits from mmap_flags
Regular mmap should not set hugepage bits for mmap_flags.
Hugepages should be addressed in mmaphuge instead.

This patch removes hugepage bits from mmap_flags to fix memory
issues with regular mmap stressor when Hugepages are allocated
but not available to containers or pods that don't have access
to the hugepages mount points.

Closes #196