stress-mmap: handle SIGBUS for hugepage mappings
Hugepage allocations with containers or pods that don't have
access to hugepages can throw a SIGBUS when access, so add
a SIGBUS handler that retries another with a new random flag
setting on the next mmap. This allows us to exercise this
corner case without breaking the stressor.

Kudos to Rafael Folco for reporting and debugging this issue
and suggesting an initial fix.

Closes ColinIanKing#196

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>