stress-gpu: Add GPU stressor
Add gpu stressor. This stressor uses OpenGL ES 2.0 and GBM API to initializes an
openGL context for stressing the gpu. We use these APIs due to the fact that
it is available in a wide range of platforms (from embedded to desktop).
The shader core usage per pixel and memory bandwidth can be modified by the gpu options.
The default frame buffer sizes x,y is 256 pixels, we can also specify options
for the stressor to perform batches of texture transfers.

Signed-off-by: Dorinda Bassey <dbassey@redhat.com>