stress-*: report file system type in file based failure messages
It is useful to report the file system type when file based
stressors fail, so add this to the error messages. If the file
system type can't be determined an empty string is added to the
end of the error message, otherwise the file system type or
file system magic type is reported.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>