stress-copy-file: handle EOF reads correctly
When starting many instances of the stressor with a very short run
time the output writes may be short on an EINTR from the alarm causing
the verify to hit end of file showing up as short reads and the verify
fails. Treat the eof condition as a successful completion to avoid this
false positive failure mode.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>