stress-cacheline: share cache data among all stressors
On systems with many instances the stresor currently does not scale well
because we have 64 stressors per instance and this soon creates context
switch overhead.  A better way is to run two processes per instance so
that two bytes next to each other in a cache line are exercised. We can
scale up to as maximum number of instances allowed and consume multiple
cache lines between them while keeping context switching down.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>