stress-plugin: add --plugin stressor to allow user code stressing via shared libs
Add --plugin stressor, allows one to load in shared library code to exercise.

The shared library can contain one or more stressor functions prefixed
with stress_ in their name. By default the plugin stressor will find
all functions prefixed with stress_ in their name and exercise these one
by one in a round-robin loop, but a specific stressor can be selected
using the --plugin-method option. The stressor function takes no
parameters and returns 0 for success and non-zero for failure (and will
terminate the plugin stressor). Each time a stressor function is executed
the bogo-op counter is incremented by one. The following example performs
10,000 nop instructions per bogo-op:

    int stress_example(void)
    {
            int i;

            for (i = 0; i < 10000; i++) {
                    __volatile__ __asm__("nop");
            }
            return 0;  /* Success */
    }

and compile the source into a shared library as, for example:

    gcc -fpic -shared -o example.so example.c

and run as using:

    stress-ng --plugin 1 --plugin-so ./example.so

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>