stress-exec: re-work per-process context and stack allocation, add vfork
To speed up process creation we pre-allocate the per process context
in a mmap'd contiguous cache. When a process is finished try and reap
it early rather than once a whole round of processes are created.
For reap'd processes, re-cycle the pre-allocated per process context
on a hot free list to use these as soon as possible once free'd.

Also re-cycle clone'd stacks, don't unmap them during the run cycle
but only unmap them at the end of the run.

By managing the process context info without any heap we get a speed
up. Also this re-working of memory allocation allows us to squeeze
in a simple vfork + execve call with minimal stack use and minimal
risk of parent heap/stack smashing.

Also add in the --exec-fork-method to the man page

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>