stress-sysinval: fix fstat/fstat64 compat check
It seems that __NR_fstat64 may be defined while __NR_fstat may be
undefined.
Looking at the kernel file 'include/uapi/asm-generic/unistd.h' this looks
possible where there is a 32-bit arch.

In any case, checking that __NR_fstat64 exists/has-a-value when adding it
to the table looks like a good idea, vs adding __NR_fstat.

This was found when updating stress-ng here:
  https://github.com/openwrt/packages/pull/19048

The ARCHS (or ARC) build is failing. No idea why it's the only one.
But this patch fixes it.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>