stress-ng: ensure checksum padding is zero and run_ok is set
The checksum data sometimes gets clobbered on some systems (e.g. some
older BSD kernels) with the bad-altstack stressor and the padding and
run_ok fields may get corrupted. Since the padding is being checksum'd
ensure it is zero'd and also ensure the run_ok flag is written back
on both true / false settings.

I'd prefer the checksum data to be in a write protected page per
stressor, but with thousands of stressors this eats too much memory
so closely packed checksums will have to do for now.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>